Make FrameDataInfo public(NFC)
